Add files via upload
The data is publicly available and shared within the works cited section.  Please download the data for yourself and create meaningful value, while respecting the investment of countless hours, training, education, time, and processes used to develop this publication as a digital product.  Thank you for taking time to read, learn, and hopefully inspiration to contribute towards a better future for all of us.