Add files via upload
Energy, Environment, and Building Systems
Exploration Using Data Science and Analytics
This document was initially produced as a series of smaller segments of analysis, then compiled to create a holistic understanding and a solution for a much more complex concern. Some of these sections have been referenced from previously researched topics, and can now be connected using data science and software programming. The objective of this publication seeks to address: (1) Health and Environmental risks from Increased Greenhouse Gas (GHG) Emissions, (2)Industry Trends and Responses, (3) Analyzing Electrical Use, (4) Technology and Communication methods within Infrastructure, (5) Analyzing a sub-segment of opportunity within Infrastructure and Energy (sensors), (6) Using Data Science and Machine Learning with this data to discovery insights, (7) Quantifying savings with rough order of magnitude estimates for energy and environmental change
The .html version of this document may also be accessed from:
https://rpubs.com/charlesah/EnergyEnvironmentSystems

This document was initially produced as a series of smaller segments of analysis, some of which have been included within this document